+++ plugins/runner/src/main/java/runner/process/DotnetProcessRunner.java
@@ -1,0 +1,11 @@
+package runner.process;
+
+public abstract class DotnetProcessRunner extends ProcessRunner {
+
+    private static final String DOTNET = "dotnet";
+
+    @Override
+    protected String getProcessName() {
+        return DOTNET;
+    }
+}
