+++ plugins/runner/src/main/java/runner/process/PythonProcessRunner.java
@@ -1,0 +1,11 @@
+package runner.process;
+
+public abstract class PythonProcessRunner extends ProcessRunner {
+
+    private static final String PYTHON = "python";
+
+    @Override
+    protected String getProcessName() {
+        return PYTHON;
+    }
+}
